jungjjeong | feat: routing 구조 변경
