+++ src/routes/index.tsx
@@ -1,4 +1,8 @@
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import {
+  createBrowserRouter,
+  RouteObject,
+  RouterProvider,
+} from "react-router-dom";
 
 import { Archive } from "@/app/archive";
 import { CommentMainPage } from "@/app/comment";
@@ -17,82 +21,117 @@ import { Done } from "@/app/post/done";
 import { KeyWord } from "@/app/post/keyword";
 import { KakaoAuth } from "@/components/app/login/kakao/kakao-auth";
 import { GlobalLayout } from "@/components/layout";
+import { RequireLoginLayout } from "@/components/layout/login-layout";
+
+type RouteChildren = {
+  auth: boolean;
+} & Omit<RouteObject, "type">;
+
+const routeChildren: RouteChildren[] = [
+  {
+    path: "/",
+    element: <OnBoarding />,
+    auth: false,
+  },
+  {
+    path: "/signup",
+    element: <SetNickName />,
+    auth: true,
+  },
+  {
+    path: "/main",
+    element: <Home />,
+    auth: true,
+  },
+  {
+    path: "/auth",
+    element: <KakaoAuth />,
+    auth: false,
+  },
+  {
+    path: "/post/write",
+    element: <Post />,
+    auth: true,
+  },
+  {
+    path: "/post/done",
+    element: <Done />,
+    auth: true,
+  },
+  {
+    path: "/post/keyword",
+    element: <KeyWord />,
+    auth: true,
+  },
+  {
+    path: "seal/:postId",
+    element: <DetailPage />,
+    auth: true,
+  },
+
+  {
+    path: "/:path",
+    element: <OnBoardingClap />,
+    auth: false,
+  },
+  {
+    path: "/clap",
+    element: <CommentMainPage />,
+    auth: false,
+  },
+  {
+    path: "/clap/write",
+    element: <CommentFormPage />,
+    auth: false,
+  },
+  {
+    path: "/clap/up",
+    element: <CommentUpPage />,
+    auth: false,
+  },
+  {
+    path: "/archive",
+    element: <Archive />,
+    auth: true,
+  },
+  {
+    path: "/mypage",
+    element: <MyPage />,
+    auth: true,
+  },
+  // {
+  //   path: "/mypage/claps",
+  //   auth: true,
+  // },
+  {
+    path: "/mypage/edit",
+    element: <MyPageEdit />,
+    auth: true,
+  },
+  {
+    path: "*",
+    element: <NotFound />,
+    auth: false,
+  },
+];
+
+const browserRouter = routeChildren.map(({ path, element, auth }) => {
+  return {
+    path,
+    element: auth ? (
+      <RequireLoginLayout>{element}</RequireLoginLayout>
+    ) : (
+      element
+    ),
+  };
+});
 
 const router = createBrowserRouter([
   {
     path: "/",
     element: <GlobalLayout />,
     errorElement: <NotFound />,
-    children: [
-      {
-        path: "/",
-        element: <OnBoarding />,
-      },
-      {
-        path: "/signup",
-        element: <SetNickName />,
-      },
-      {
-        path: "/main",
-        element: <Home />,
-      },
-      {
-        path: "/auth",
-        element: <KakaoAuth />,
-      },
-      {
-        path: "/post/write",
-        element: <Post />,
-      },
-      {
-        path: "/post/done",
-        element: <Done />,
-      },
-      {
-        path: "/post/keyword",
-        element: <KeyWord />,
-      },
-      {
-        path: "seal/:postId",
-        element: <DetailPage />,
-      },
-
-      {
-        path: "/:path",
-        element: <OnBoardingClap />,
-      },
-      {
-        path: "/clap",
-        element: <CommentMainPage />,
-      },
-      {
-        path: "/clap/write",
-        element: <CommentFormPage />,
-      },
-      {
-        path: "/clap/up",
-        element: <CommentUpPage />,
-      },
-      {
-        path: "/archive",
-        element: <Archive />,
-      },
-      {
-        path: "/mypage",
-        element: <MyPage />,
-      },
-      {
-        path: "/mypage/claps",
-      },
-      {
-        path: "/mypage/edit",
-        element: <MyPageEdit />,
-      },
-      {
-        path: "*",
-        element: <NotFound />,
-      },
-    ],
+    children: browserRouter,
   },
 ]);
